adds themeContext module
