+++ package.json
@@ -5,6 +5,7 @@
   "keywords": [],
   "main": "src/index.js",
   "dependencies": {
+    "@blueprintjs/core": "^3.53.0",
     "node-sass": "^6.0.1",
     "react": "17.0.2",
     "react-bootstrap": "^1.6.1",
@@ -14,7 +15,6 @@
     "react-scripts": "4.0.3",
     "uuid": "^8.3.2"
   },
-  "devDependencies": {},
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
