--- src/components/todo/header.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-function Header(props) {
-  return (
-    <h1>To Do List: {props.incomplete} items pending</h1>
-  )
-}
-
-export default Header;
