--- src/components/todo/list.js
@@ -1,30 +1,0 @@
-import React from "react";
-
-
-function List(props) {
-  return (
-    props.list.map(item => (
-      <div key={item.id}>
-        <p>{item.text}</p>
-        <p><small>Assigned to: {item.assignee}</small></p>
-        <p><small>Difficulty: {item.difficulty}</small></p>
-        <div onClick={() => props.toggleComplete(item.id)}>Complete: {item.complete.toString()}</div>
-        <hr />
-      </div>
-    ))
-  )
-}
-
-export default List;
-
-// {
-//   list.map(item => (
-//     <div key={item.id}>
-//       <p>{item.text}</p>
-//       <p><small>Assigned to: {item.assignee}</small></p>
-//       <p><small>Difficulty: {item.difficulty}</small></p>
-//       <div onClick={() => toggleComplete(item.id)}>Complete: {item.complete.toString()}</div>
-//       <hr />
-//     </div>
-//   ))
-// }
