+++ src/components/todo/todo.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import useForm from '../../hooks/form.js';
-import Header from './header.js';
-import Form from './form.js';
-import List from './list.js';
+import Header from '../Header/header.js';
+import Form from '../Form/form.js';
+import List from '../List/list.js';
 
 import { v4 as uuid } from 'uuid';
 
@@ -45,51 +45,13 @@ const ToDo = () => {
 
   return (
     <>
+
       <Header incomplete={incomplete} />
-      {/* <header>
-        <h1>To Do List: {incomplete} items pending</h1>
-      </header> */}
 
       <Form handleChange={handleChange} handleSubmit={handleSubmit} />
 
-      {/* <form onSubmit={handleSubmit}>
-
-        <h2>Add To Do Item</h2>
-
-        <label>
-          <span>To Do Item</span>
-          <input onChange={handleChange} name="text" type="text" placeholder="Item Details" />
-        </label>
-
-        <label>
-          <span>Assigned To</span>
-          <input onChange={handleChange} name="assignee" type="text" placeholder="Assignee Name" />
-        </label>
-
-        <label>
-          <span>Difficulty</span>
-          <input onChange={handleChange} defaultValue={3} type="range" min={1} max={5} name="difficulty" />
-        </label>
-
-        <label>
-          <button type="submit">Add Item</button>
-        </label>
-      </form> */}
-
       <List list={list} toggleComplete={toggleComplete} />
-
-      {/* {list.map(item => (
-        <div key={item.id}>
-          <p>{item.text}</p>
-          <p><small>Assigned to: {item.assignee}</small></p>
-          <p><small>Difficulty: {item.difficulty}</small></p>
-          <div onClick={() => toggleComplete(item.id)}>Complete: {item.complete.toString()}</div>
-          <hr />
-        </div>
-      ))} */}
-
     </>
-
   );
 };
